Code updated

+++ src/main/java/com/prescription/prescriptioncreator/model/PreviousHistoryDetails.java
@@ -13,13 +13,14 @@ public class PreviousHistoryDetails {
     @Override
     public String toString() {return this.getPrevious_history();
     }
-    ///
-    public int getId() {
+
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
-    private int id;
+
+    private long id;
 }
